services/img-recon/requirements.txt
json-rpc
requests
prettyconf
werkzeug
numpy
opencv-python
python-slugify
gunicorn